exec: Fix incr/decr for booleans
Pushed increment and decrement to the runtime so they could do the
proper type checks.

Also made the count a u32 - didn't see a point in having it be
platform-dependent, or any larger.

+++ src/ast.rs
@@ -14,8 +14,8 @@ pub struct Statement {
 #[derive(Debug)]
 pub enum StatementKind {
     Assign(LValue, Expr),
-    Incr(LValue, usize),
-    Decr(LValue, usize),
+    Incr(LValue, u32),
+    Decr(LValue, u32),
     Say(Expr),
 
     Continue,
